Fix color mappings for solarized-dark.reg
All of the color pairings (e.g. `blue` and `brblue`) were swapped
incorrectly, leading to the `blue` console color to be `base0`,
instead of `brblue`, the intended mapping.

Using the 16-bit terminal color notation, the following color pairs
were exchanged:

- `brback` and `black`
- `brgreen` and `green`
- `bryellow` and `yellow`
- `brblue` and `blue`
- `brcyan` and `cyan`
- `brwhite` and `white`
- `brred` and `red`
- `brmagenta` and `magenta`

Because the indices changed, the screen and popup color pairs had to
be updated too.

This fixes the colors in the (new) Windows 10 console (used by the
Windows Subsystem for Linux AKA Bash on Ubuntu on Windows, and
PowerShell).