Fix foreground colors in Set-SolarizedDarkColorDefaults
The `CommentForegroundColor` was added to be explicit, though it was
already correctly `base01` AKA DarkGreen.

The `EmphasisForegroundColor` was fixed to be `base1` AKA DarkCyan.

The "colorful" foreground colors were fixed to be their non-dark
versions, that is, those that are colorful (not repurposed to the base
tones of Solarized). So the `MemberForegroundColor` is now `cyan` aka
Cyan, instead of `base1` AKA DarkCyan, etc.