Add option to use Unix Solarized color assignments
Adds a 'Shell' parameter to Update-Link. It takes two values 'Windows'
is the default and maintains the existing behaviour. 'Unix' assigns
Solarized colors to the same ANSI color escape codes as the terminal
emulator themes linked to from https://ethanschoonover.com/solarized/.

The Unix option allows Solarized themes for Unix tools like
https://github.com/seebi/dircolors-solarized work correctly and is
designed to be used with Cygwin or WSL.